requirements.txt
torch==2.2.2
torchvision
torchaudio
diffusers==0.31.0
transformers==4.41.2
accelerate
safetensors==0.4.3
gTTS
moviepy
pillow
nltk
